Fifth addition:
Added a section for "random_card" which creates a shuffled deck of cards to then display them at random, resetting when the deck runs out or is told to do so.

This is done by creating a print_card() function to avoid repeat code for each game, then adding a create_shuffled_ideck and shuffle_ideck functions. The ideck refers to an index deck for the model deck which remains ordered - avoiding shuffling a deck of complicated structs. The memory for the ideck is, as with the model deck, manually allocated. The shuffled deck is then worked through, using the model_deck as a printout.

The advantage of this approach is you won't get the same card twice unless shuffled again (like dealing it normally) giving it a realistic conditional probability vibe. This is compared to a random number generator which has an equal chance of all cards regardless of previous rounds, this is relevant to say pontoon where you can count cards.

Next step is replace the embedded if statements with a switch statement for efficiency.